+++ packages/terra-data-grid/src/index.js
@@ -1,0 +1,5 @@
+import WorklistDataGrid from './WorklistDataGrid';
+
+// TODO: Re-enable linter rule when FlowsheetDataGrid is added.
+// eslint-disable-next-line import/prefer-default-export
+export { WorklistDataGrid };