+++ packages/terra-framework-docs/src/terra-dev-site/doc/data-grid/WorklistDataGrid.1/Examples.3/DefaultWorklistDataGrid.jsx
@@ -1,15 +1,13 @@
-import React, { useState } from 'react';
-import WorklistDataGrid from 'terra-worklist-data-grid';
+import React from 'react';
+import { WorklistDataGrid } from 'terra-data-grid';
 
 const gridDataJSON = {
   cols: [
-    {
-      id: 'Column-0', displayName: 'Patient', sortIndicator: 'ascending', hasError: true,
-    },
+    { id: 'Column-0', displayName: 'Patient' },
     { id: 'Column-1', displayName: 'Location' },
     { id: 'Column-2', displayName: 'Illness Severity' },
     { id: 'Column-3', displayName: 'Visit' },
-    { id: 'Column-4', displayName: 'Allergy', hasError: true },
+    { id: 'Column-4', displayName: 'Allergy' },
     { id: 'Column-5', displayName: 'Primary Contact' },
     { id: 'Column-6', displayName: 'Generic Order Counts' },
     { id: 'Column-7', displayName: 'Patient Age' },
@@ -57,48 +55,12 @@ const DefaultWorklistDataGrid = () => {
   const rowHeaderIndex = 0;
   const { cols, rows } = gridDataJSON;
 
-  const [gridColumns, setGridColumns] = useState(cols);
-  const [gridRows, setGridRows] = useState(rows);
-
-  const onColumnSelect = (columnId) => {
-    const newColumnArray = gridColumns.map((column, columnIndex) => {
-      const newColumn = { ...column };
-
-      if (column.id === columnId) {
-        newColumn.sortIndicator = column.sortIndicator === 'ascending' ? 'descending' : 'ascending';
-
-        const newGridRows = [...gridRows];
-        newGridRows.sort((rowA, rowB) => {
-          const firstRowContent = rowA.cells[columnIndex].content || '';
-          const secondRowContent = rowB.cells[columnIndex].content || '';
-
-          if (newColumn.sortIndicator === 'ascending') {
-            return firstRowContent.localeCompare(secondRowContent);
-          }
-          return secondRowContent.localeCompare(firstRowContent);
-        });
-
-        setGridRows(newGridRows);
-      } else {
-        newColumn.sortIndicator = undefined;
-      }
-
-      return newColumn;
-    });
-
-    setGridColumns(newColumnArray);
-  };
-
   return (
     <WorklistDataGrid
       id="default-terra-worklist-data-grid"
-      overflowColumns={gridColumns}
-      rows={gridRows}
+      overflowColumns={cols}
+      rows={rows}
       rowHeaderIndex={rowHeaderIndex}
-      rowHeight="50px"
-      defaultColumnWidth={100}
-      columnHeaderHeight="50px"
-      onColumnSelect={onColumnSelect}
       ariaLabel="Worklist Data Grid"
     />
   );