+++ packages/terra-framework-docs/src/terra-dev-site/test/data-grid/worklist-data-grid/WorklistDataGridFocusableCell.test.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import WorklistDataGrid from 'terra-worklist-data-grid';
+import { WorklistDataGrid } from 'terra-data-grid';
 import NotificationDialog from 'terra-notification-dialog';
 
 const WorklistDataGridFocusableCell = () => {